+++ R/mod_compare_nth_cases_plot.R
@@ -7,6 +7,7 @@
 #' @param actives if TRUE then add new_active and active variables to vars.
 #' @param tests if TRUE then add new_test and test variables to vars.
 #' @param hosp if TRUE then add new_hosp and hosp variables to vars.
+#' @param selectvar character variable selected in ui.
 #'
 #' @noRd
 #'
@@ -17,7 +18,7 @@ mod_compare_nth_cases_plot_ui <- function(id, vars = c("confirmed", "deaths", "r
                                                        "new_prevalence_rate_1M_pop",
                                                        "new_tests", "new_tests_rate_1M_pop","new_positive_tests_rate",
                                                         "growth_factor_3", "lethality_rate" ),
-                                          actives = TRUE, tests = FALSE, hosp = FALSE){
+                                          actives = TRUE, tests = FALSE, hosp = FALSE, selectvar = "new_confirmed"){
   ns <- NS(id)
 
   choices_plot = varsNames(vars)
@@ -47,7 +48,7 @@ mod_compare_nth_cases_plot_ui <- function(id, vars = c("confirmed", "deaths", "r
       column(7,
              offset = 1,
              selectInput(inputId = ns("radio_indicator"), label = "",
-                          choices = choices_plot, selected ="new_confirmed")
+                          choices = choices_plot, selected = selectvar)
       ),
       column(4,
              selectInput(inputId = ns("radio_log_linear"), label = "",
@@ -62,7 +63,7 @@ mod_compare_nth_cases_plot_ui <- function(id, vars = c("confirmed", "deaths", "r
 #' compare_nth_cases_plot Server Function
 #'
 #' @param df data.frame
-#' @param n min number of cases for a country to be considered. Default 1000
+#' @param nn minimum date derived from first day with more than nn cases. Default 1000
 #' @param w number of days of outbreak. Default 7
 #' @param n_highligth number of countries to highlight if istop == TRUE
 #' @param istop logical to choose title, if top n_highligth countries are selected
@@ -79,12 +80,14 @@ mod_compare_nth_cases_plot_ui <- function(id, vars = c("confirmed", "deaths", "r
 #'
 #' @noRd
 mod_compare_nth_cases_plot_server <- function(input, output, session, df,
-                                              n = 1000, w = 7,
-                                              n_highligth = 5, istop = TRUE, g_palette = graph_palette){
+                                              nn = 1000, w = 7,
+                                              n_highligth = 5, istop = TRUE, g_palette = graph_palette, datevar = "date"){
   ns <- session$ns
+  df$Date = df[[datevar]]
 
   # Give DF standard structure; reacts to input$radio_indicator
   df_data <- reactive({
+
     if(istop) {
       countries_order =  df %>% filter(date == max(date)) %>%
         arrange(desc(!!as.symbol(req(input$radio_indicator)))) %>%
@@ -95,14 +98,20 @@ mod_compare_nth_cases_plot_server <- function(input, output, session, df,
     } else {
       data = df
     }
-    df_tmp <- data %>% .[,c("Country.Region", req(input$radio_indicator), "contagion_day")] %>%
+
+    # filter off x before nn
+    date_first_contagion = min(data$date[data$confirmed >= nn])
+    data = data[data$date >= date_first_contagion, ]
+
+    df_tmp <- data %>% .[,c("Country.Region", req(input$radio_indicator), "Date")] %>%
       bind_cols(data[,req(input$radio_indicator)] %>% setNames("Value")) %>%
       rename(Status = Country.Region ) %>%
-      rename(Date = contagion_day ) %>%
+      #rename(Date = contagion_day ) %>%
       select(-req(input$radio_indicator))
 
+    # filter dates with 0 contagions
 
-    if (istop && ("China" %in% df_tmp$Status)) {
+    if (istop && ("China" %in% df_tmp$Status) && datevar == "contagion_day") {
       # Day of the country with max contagions after china
       max_contagion_no_china <- df_tmp %>%
         filter(Status != "China") %>%
@@ -125,7 +134,8 @@ mod_compare_nth_cases_plot_server <- function(input, output, session, df,
 
   # Plot -----
   output$plot <- renderPlotly({
-    p <- plot_all_highlight(df_data(), log = log(), text = "Area", n_highligth = n_highligth, percent = ifelse(req(input$radio_indicator) %in% rate_vars, TRUE, FALSE), date_x = FALSE, g_palette)
+    p <- plot_all_highlight(df_data(), log = log(), text = "Area", n_highligth = n_highligth, percent = ifelse(req(input$radio_indicator) %in% rate_vars, TRUE, FALSE),
+                            date_x = ifelse(datevar == "date", TRUE,FALSE), g_palette)
     p <- p %>%
       plotly::ggplotly(tooltip = c("text", "x_tooltip", "y_tooltip")) %>%
       plotly::layout(legend = list(orientation = "h", y = 1.1, yanchor = "bottom", itemsizing = "constant"))
@@ -135,16 +145,18 @@ mod_compare_nth_cases_plot_server <- function(input, output, session, df,
 
   if (istop) {
     output$title <- renderUI({
-      div(h4(paste0("Top ",n_highligth," countries from day of ", n ," contagion")), align = "center", style = "margin-top:20px; margin-bottom:20px;")
+      div(h4(paste0("Top ",n_highligth," countries from day of ", nn ," contagion")), align = "center", style = "margin-top:20px; margin-bottom:20px;")
     })
   } else {
     output$title <- renderUI({
-      div(h4(paste0("Timeline from day of ", n ," contagion")), align = "center", style = "margin-top:20px; margin-bottom:20px;")
+      div(h4(paste0("Timeline from day with ", nn ," contagions")), align = "center", style = "margin-top:20px; margin-bottom:20px;")
     })
   }
 
   output$caption <- renderUI({
-      p(paste0("Computed as rolling weekly average. Day 0 is the day when ", n," confirmed cases are reached."))
+
+      p(paste0("Computed as rolling weekly average. ",ifelse(datevar == "date", "First day", "Contagion day 0"),
+               " is the day when ", nn," confirmed cases are reached."))
   })
 }